+++ components/generator/array/arrayGenerator.js
@@ -7,8 +7,8 @@ import {
   FormControlLabel,
   Checkbox,
   Radio,
-  InputLabel,
   MenuItem,
+  InputLabel,
 } from "@mui/material";
 import { CopyToClipboard } from "react-copy-to-clipboard";
 import FileCopyIcon from "@mui/icons-material/FileCopy";
@@ -18,72 +18,94 @@ import DownloadIcon from "@mui/icons-material/GetApp";
 
 import toast, { Toaster } from "react-hot-toast";
 
-import StringGeneratorFunc from "components/generatorFunc/stringGeneratorFunc";
+import ArrayGeneratorFunc from "components/generator/array/arrayGeneratorFunc";
 
 const options = [
-  "Hide Length",
-  "Hide Number of Strings",
-  "Distinct Strings",
-  "Distinct Strings (Case Insensitive)",
-  "Distinct Characters",
+  "Negative Outputs",
+  "Hide Array Size",
+  "Distinct Elements",
+  "Show Total Cases",
 ];
 
-const GenerateString = () => {
+const GenerateArray = () => {
   const {
-    stringLength,
-    setStringLength,
-    numStrings,
-    setNumStrings,
-    excludedChars,
-    setExcludedChars,
-    includedChars,
-    setIncludedChars,
-    generatedStrings,
+    minValue,
+    setMinValue,
+    maxValue,
+    setMaxValue,
+    arraySize,
+    setArraySize,
+    numArrays,
+    setNumArrays,
+    generatedValues,
     copied,
     timeTaken,
-    smallAlphabets,
-    setSmallAlphabets,
-    capitalAlphabets,
-    setCapitalAlphabets,
-    numbers,
-    setNumbers,
-    specialChars,
-    setSpecialChars,
+    isFloat,
+    setIsFloat,
     randomSize,
     setRandomSize,
     advanceOptions,
     isLoading,
-    handleGenerateStrings,
-    handleCopyStrings,
+    handleGenerateValues,
+    handleCopyValues,
     handleDownloadValues,
     handleResetValues,
     handleAdvanceOptionChange,
+    handleSortChange,
+    handleOptionChange,
+    any,
+    odd,
+    even,
+    prime,
     increasing,
     decreasing,
     random,
-    handleSortChange,
-  } = StringGeneratorFunc();
+  } = ArrayGeneratorFunc();
+
+  console.log(isFloat);
 
   return (
     <StyledComponents.StyledGrid container>
       <Toaster />
       <Grid item xs={12} sm={8} md={8} sx={{ margin: "auto" }}>
         <StyledComponents.StyledCard>
-          <StyledComponents.StyledCardHeader title="Random String Generator" />
+          <StyledComponents.StyledCardHeader title="Generate Array" />
           <StyledComponents.StyledCardContent>
             <Grid container spacing={2}>
               <Grid item xs={6}>
-                <Tooltip title="Enter the number of string">
+                <Tooltip title="Enter the minimum value for the array">
                   <StyledComponents.StyledTextField
-                    label="Number of Strings"
+                    label="Min Value"
                     type="number"
-                    value={numStrings}
+                    value={minValue}
+                    onChange={(e) => setMinValue(e.target.value)}
+                    fullWidth
+                  />
+                </Tooltip>
+              </Grid>
+              <Grid item xs={6}>
+                <Tooltip title="Enter the maximum value for the array">
+                  <StyledComponents.StyledTextField
+                    label="Max Value"
+                    type="number"
+                    value={maxValue}
+                    onChange={(e) => setMaxValue(e.target.value)}
+                    fullWidth
+                  />
+                </Tooltip>
+              </Grid>
+              <Grid item xs={6}>
+                <Tooltip title="Enter the size of the array">
+                  <StyledComponents.StyledTextField
+                    label="Array Size"
+                    type="number"
+                    value={arraySize}
                     onChange={(e) => {
                       if (e.target.value < 0) {
                         toast.error("Please enter a positive number");
-                        setNumStrings(1);
+                        setArraySize(10);
                       } else {
-                        setNumStrings(e.target.value);
+                        setArraySize(e.target.value);
                       }
                     }}
                     fullWidth
@@ -91,17 +113,17 @@ const GenerateString = () => {
                 </Tooltip>
               </Grid>
               <Grid item xs={6}>
-                <Tooltip title="Enter the length of the string">
+                <Tooltip title="Enter the number of arrays to generate">
                   <StyledComponents.StyledTextField
-                    label="string length"
+                    label="Number of Arrays"
                     type="number"
-                    value={stringLength}
+                    value={numArrays}
                     onChange={(e) => {
                       if (e.target.value < 0) {
                         toast.error("Please enter a positive number");
-                        setStringLength(10);
+                        setNumArrays(1);
                       } else {
-                        setStringLength(e.target.value);
+                        setNumArrays(e.target.value);
                       }
                     }}
                     fullWidth
@@ -110,8 +132,22 @@ const GenerateString = () => {
               </Grid>
             </Grid>
             <Grid container spacing={2} sx={{ marginTop: "1rem" }}>
+              <Grid item xs={3}>
+                <Tooltip title="Check to generate float values">
+                  <FormControlLabel
+                    control={
+                      <Checkbox
+                        checked={isFloat}
+                        onChange={(e) => setIsFloat(e.target.checked)}
+                        name="isFloat"
+                      />
+                    }
+                    label="Generate Float Arrays"
+                  />
+                </Tooltip>
+              </Grid>
               <Grid item xs={2}>
-                <Tooltip title="Check for random size">
+                <Tooltip title="Check to generate random size arry">
                   <FormControlLabel
                     control={
                       <Checkbox
@@ -144,89 +180,74 @@ const GenerateString = () => {
               </Grid>
             </Grid>
             <Grid container spacing={2} sx={{ marginTop: "1rem" }}>
-              <Grid item xs={3}>
-                <Tooltip title="Check to include a-z">
+              <Grid item xs={2}>
+                <Tooltip title="Check to generate any values">
                   <FormControlLabel
                     control={
-                      <Checkbox
-                        checked={smallAlphabets}
-                        onChange={(e) => setSmallAlphabets(e.target.checked)}
-                        name="a-z"
+                      <Radio
+                        checked={any}
+                        onChange={handleOptionChange}
+                        value="any"
+                        name="radio-button-demo"
+                        inputProps={{ "aria-label": "any" }}
                       />
                     }
-                    label="a-z"
+                    label="any"
                   />
                 </Tooltip>
               </Grid>
-              <Grid item xs={3}>
-                <Tooltip title="Check to include A-Z">
+              <Grid item xs={2}>
+                <Tooltip title="Check to generate even values">
                   <FormControlLabel
                     control={
-                      <Checkbox
-                        checked={capitalAlphabets}
-                        onChange={(e) => setCapitalAlphabets(e.target.checked)}
-                        name="A-Z"
+                      <Radio
+                        checked={even}
+                        onChange={handleOptionChange}
+                        value="even"
+                        name="radio-button-demo"
+                        inputProps={{ "aria-label": "even" }}
                       />
                     }
-                    label="A-Z"
+                    label="Even"
                   />
                 </Tooltip>
               </Grid>
-              <Grid item xs={3}>
-                <Tooltip title="Check to include 0-9">
+              <Grid item xs={2}>
+                <Tooltip title="Check to generate odd values">
                   <FormControlLabel
                     control={
-                      <Checkbox
-                        checked={numbers}
-                        onChange={(e) => setNumbers(e.target.checked)}
-                        name="0-9"
+                      <Radio
+                        checked={odd}
+                        onChange={handleOptionChange}
+                        value="odd"
+                        name="radio-button-demo"
+                        inputProps={{ "aria-label": "odd" }}
                       />
                     }
-                    label="0-9"
+                    label="Odd"
                   />
                 </Tooltip>
               </Grid>
-              <Grid item xs={3}>
-                <Tooltip title="Check to include special chars">
+              <Grid item xs={2}>
+                <Tooltip title="Check to generate prime values">
                   <FormControlLabel
                     control={
-                      <Checkbox
-                        checked={specialChars}
-                        onChange={(e) => setSpecialChars(e.target.checked)}
-                        name="special chars"
+                      <Radio
+                        checked={prime}
+                        onChange={handleOptionChange}
+                        value="prime"
+                        name="radio-button-demo"
+                        inputProps={{ "aria-label": "prime" }}
                       />
                     }
-                    label="special chars"
-                  />
-                </Tooltip>
-              </Grid>
-            </Grid>
-            <Grid container spacing={2} sx={{ marginTop: "1rem" }}>
-              <Grid item xs={6}>
-                <Tooltip title="Enter the chars to be exclude">
-                  <StyledComponents.StyledTextField
-                    label="Exclude Characters"
-                    value={excludedChars}
-                    onChange={(e) => setExcludedChars(e.target.value)}
-                    fullWidth
-                  />
-                </Tooltip>
-              </Grid>
-
-              <Grid item xs={6}>
-                <Tooltip title="Enter the chars to be include">
-                  <StyledComponents.StyledTextField
-                    label="Include Characters"
-                    value={includedChars}
-                    onChange={(e) => setIncludedChars(e.target.value)}
-                    fullWidth
+                    label="Prime"
                   />
                 </Tooltip>
               </Grid>
             </Grid>
             <Grid container spacing={2} sx={{ marginTop: "1rem" }}>
               <Grid item xs={4}>
-                <Tooltip title="Check to generate increasing values(Only work with multiple strings)">
+                <Tooltip title="Check to generate increasing values">
                   <FormControlLabel
                     control={
                       <Radio
@@ -235,7 +256,6 @@ const GenerateString = () => {
                         value="increasing"
                         name="radio-button-demo"
                         inputProps={{ "aria-label": "increasing" }}
-                        disabled={numStrings <= 1}
                       />
                     }
                     label="Increasing"
@@ -243,7 +263,7 @@ const GenerateString = () => {
                 </Tooltip>
               </Grid>
               <Grid item xs={4}>
-                <Tooltip title="Check to generate decreasing values(Only work with multiple strings)">
+                <Tooltip title="Check to generate decreasing values">
                   <FormControlLabel
                     control={
                       <Radio
@@ -252,7 +272,6 @@ const GenerateString = () => {
                         value="decreasing"
                         name="radio-button-demo"
                         inputProps={{ "aria-label": "decreasing" }}
-                        disabled={numStrings <= 1}
                       />
                     }
                     label="Decreasing"
@@ -260,7 +279,7 @@ const GenerateString = () => {
                 </Tooltip>
               </Grid>
               <Grid item xs={4}>
-                <Tooltip title="Check to generate random values(Only work with multiple strings)">
+                <Tooltip title="Check to generate random values">
                   <FormControlLabel
                     control={
                       <Radio
@@ -269,7 +288,6 @@ const GenerateString = () => {
                         value="random"
                         name="radio-button-demo"
                         inputProps={{ "aria-label": "random" }}
-                        disabled={numStrings <= 1}
                       />
                     }
                     label="Random"
@@ -281,36 +299,33 @@ const GenerateString = () => {
               <Grid item xs={12}>
                 <StyledComponents.StyledButton
                   variant="contained"
-                  onClick={handleGenerateStrings}
-                  disabled={isLoading}
                   fullWidth
                   startIcon={<GenerateIcon />}
+                  onClick={handleGenerateValues}
+                  disabled={isLoading}
                 >
-                  Generate
+                  Generate Array
                 </StyledComponents.StyledButton>
+              </Grid>
+              <Grid item xs={6}>
+                <CopyToClipboard onCopy={handleCopyValues}>
+                  <StyledComponents.StyledButton
+                    variant="contained"
+                    fullWidth
+                    startIcon={<FileCopyIcon />}
+                    disabled={isLoading}
+                  >
+                    {copied ? "Copied" : "Copy to clipboard"}
+                  </StyledComponents.StyledButton>
+                </CopyToClipboard>
               </Grid>
               <Grid item xs={6}>
                 <StyledComponents.StyledButton
                   variant="contained"
-                  onClick={handleCopyStrings}
-                  disabled={isLoading}
-                  fullWidth
-                  startIcon={
-                    <CopyToClipboard onCopy={handleCopyStrings}>
-                      <FileCopyIcon />
-                    </CopyToClipboard>
-                  }
-                >
-                  {copied ? "Copied" : "Copy to clipboard"}
-                </StyledComponents.StyledButton>
-              </Grid>
-              <Grid item xs={6}>
-                <StyledComponents.StyledButton
-                  variant="contained"
-                  onClick={handleResetValues}
-                  disabled={isLoading}
                   fullWidth
                   startIcon={<RefreshIcon />}
+                  onClick={handleResetValues}
+                  disabled={isLoading}
                 >
                   Reset
                 </StyledComponents.StyledButton>
@@ -331,27 +346,29 @@ const GenerateString = () => {
               <Grid item xs={6}>
                 {timeTaken && <p>Time taken: {timeTaken}</p>}
                 <StyledComponents.StyledTypography variant="h6">
-                  Generated String
+                  Generated Array
                 </StyledComponents.StyledTypography>
               </Grid>
               <Grid item xs={12}>
                 <StyledComponents.StyledTypography variant="subtitle">
-                  {generatedStrings.length === 0 ? (
-                    <div>No Strings generated yet</div>
+                  {generatedValues.length === 0 ? (
+                    <div>No Array generated yet</div>
                   ) : null}
-                  {generatedStrings.length > 0 && (
-                    <Typography variant="subtitle">
-                      {!advanceOptions.includes("Hide Number of Strings") &&
-                        generatedStrings.length}
-                      {generatedStrings.map((str, index) => (
-                        <div key={index}>
-                          {!advanceOptions.includes("Hide Length") && (
-                            <div>{str.length}</div>
-                          )}
-                          {str}
-                        </div>
-                      ))}
-                    </Typography>
+                  {generatedValues.length > 0 && (
+                    <>
+                      <Typography variant="subtitle">
+                        {advanceOptions.includes("Show Total Cases") &&
+                          generatedValues.length}
+                        {generatedValues.map((array, index) => (
+                          <div key={index}>
+                            {!advanceOptions.includes("Hide Array Size") && (
+                              <div>{array.length}</div>
+                            )}
+                            {array.join(", ")}
+                          </div>
+                        ))}
+                      </Typography>
+                    </>
                   )}
                 </StyledComponents.StyledTypography>
               </Grid>
@@ -363,4 +380,4 @@ const GenerateString = () => {
   );
 };
 
-export default GenerateString;
+export default GenerateArray;
