+++ components/generator/array/arrayGeneratorFunc.js
