+++ components/generator/graph/graphGeneratorFunc.js
