+++ components/generator/integer/integerGeneratorFunc.js
