+++ components/generator/linkedList/linkedListGeneratorFunc.js
