+++ components/generator/palindrome/palindromeGeneratorFunc.js
