+++ components/generator/string/stringGeneratorFunc.js
