+++ pages/generator/arrayGenerator.js
@@ -1,4 +1,4 @@
-import GenerateArray from "@/components/generator/arrayGenerator";
+import GenerateArray from "@/components/generator/array/arrayGenerator";
 
 export default function ArrayGenerator() {
   return <GenerateArray />;
