+++ pages/generator/integerGenerator.js
@@ -1,4 +1,4 @@
-import GenerateInteger from "@/components/generator/integerGenerator";
+import GenerateInteger from "@/components/generator/integer/integerGenerator";
 
 export default function IntegerGenerator() {
   return <GenerateInteger />;
