+++ pages/generator/linkedListGenerator.js
@@ -1,4 +1,4 @@
-import GenerateLinkedList from "@/components/generator/linkedListGenerator";
+import GenerateLinkedList from "@/components/generator/linkedList/linkedListGenerator";
 
 export default function LinkedListGenerator() {
   return <GenerateLinkedList />;
