+++ pages/generator/palindromeGenerator.js
@@ -1,4 +1,4 @@
-import PalindromeGenerator from "@/components/generator/palindromeGenerator";
+import PalindromeGenerator from "@/components/generator/palindrome/palindromeGenerator";
 
 export default function PalindromeGeneratorPage() {
   return <PalindromeGenerator />;
