+++ pages/generator/stringGenerator.js
@@ -1,4 +1,4 @@
-import GenerateString from "@/components/generator/stringGenerator";
+import GenerateString from "@/components/generator/string/stringGenerator";
 
 export default function StringGenerator() {
   return <GenerateString />;
